requirements.txt
setuptools<81
python-dotenv
lark-oapi
aiohttp
flask
waitress
redis
gunicorn